backend/pyproject.toml
[tool.poetry]
name = "kids-pixel-pals-backend"
version = "0.1.0"
description = "Backend API for Kids Pixel Pals - Safe social platform for children"
authors = ["Kids Pixel Pals Team"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
cryptography = "^41.0.7"
argon2-cffi = "^23.1.0"
redis = "^5.0.1"
python-socketio = {extras = ["asgi"], version = "^5.10.0"}
aiohttp = "^3.9.1"
httpx = "^0.25.2"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"